feat: ProtectedRoute AuthChecker
